+++ js/data.js
@@ -1,0 +1,41 @@
+export const DESCRIPTIONS = [
+  'Новогодняя ель',
+  'Вечеринка',
+  'Новый автомобиль',
+  'Однажды вечером',
+  'Оркестр',
+  'Кроссовки',
+  'Крабик',
+  'Ретро автомобиль',
+  'Пролетая над Африкой',
+  'Легкий перекус',
+  'Котик',
+  'Отель на пляже',
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Женя Пупсик',
+  'Котик',
+  'Саша Иванова',
+  'Просто Вася',
+  'Лео',
+  'Фунтик',
+  'Катерина',
+  'Петр Сидоренко',
+  'Линда',
+  'Джек',
+  'Людмила',
+  'Федя Федоров',
+  'Матроскин',
+  'Барбоскин',
+  'Павел Первый',
+];
